Fixes admin email overwrite bug in sendEmail.
The checks for per-host warning messages overwrite the configuration
with that of the individual hosts. As a result, the email regarding bad
hosts is sent to the EMailAdminUserName of the last host that is
checked. This change fixes the configuration overwrite and sends
separate emails regarding bad hosts to their respective admins.